requirements.txt
beautifulsoup4==4.12.3
pydantic==2.6.4
PyYAML==6.0.1
Requests==2.31.0
selenium==4.18.1
